+++ react-components/src/components/CostsAndPricing/constants.js
@@ -3,8 +3,11 @@ export const direct = [
     label: 'Product cost',
     id: 'product_costs',
     placeholder: '0',
-    tooltip:
-      '<h3>What are product costs?</h3><p>These are the costs of the parts, materials or ingredients that go onto making the end product.</p><p>This does not include labour costs.</p>',
+    tooltip: {
+      content:
+        '<p>These are the costs of the parts, materials or ingredients that go onto making the end product.</p><p>This does not include labour costs.</p>',
+      title: 'What are product costs?',
+    },
     type: 'number',
     field: 'direct_costs',
   },
@@ -12,8 +15,11 @@ export const direct = [
     label: 'Labour cost',
     id: 'labour_costs',
     placeholder: '0',
-    tooltip:
-      '<h3>What are labour costs?</h3><p>These are the costs related to paying wages for employees who work on making your product directly including workers on your assembly line.</p>',
+    tooltip: {
+      content:
+        '<p>These are the costs related to paying wages for employees who work on making your product directly including workers on your assembly line.</p>',
+      title: 'What are labour costs?',
+    },
     type: 'number',
     field: 'direct_costs',
   },
@@ -21,7 +27,7 @@ export const direct = [
     label: 'Additional margin',
     id: 'other_direct_costs',
     placeholder: '0',
-    tooltip: '',
+    tooltip: {},
     type: 'number',
     field: 'direct_costs',
   },
@@ -32,8 +38,11 @@ export const overhead = [
     label: 'Product adaptation',
     id: 'product_adaption',
     placeholder: '0',
-    tooltip:
-      '<h3>What is product adaptation?</h3><p>These are any changes you need to make to sell abroad. This includes translations, rebranding, packaging and labelling, and meeting local regulations.</p>',
+    tooltip: {
+      content:
+        '<p>These are any changes you need to make to sell abroad. This includes translations, rebranding, packaging and labelling, and meeting local regulations.</p>',
+      title: 'What is product adaptation?',
+    },
     type: 'number',
     field: 'overhead_costs',
   },
@@ -41,8 +50,11 @@ export const overhead = [
     label: 'Freight and logistics',
     id: 'freight_logistics',
     placeholder: '0',
-    tooltip:
-      '<h3>What is freight and logistics?</h3><p>Logistics is the process of getting your goods to their final destination.</p><p>A freight forwarder is a third-party agent that most UK companies use to transport their goods.</p>',
+    tooltip: {
+      content:
+        '<p>Logistics is the process of getting your goods to their final destination.</p><p>A freight forwarder is a third-party agent that most UK companies use to transport their goods.</p>',
+      title: 'What is freight and logistics?',
+    },
     type: 'number',
     field: 'overhead_costs',
   },
@@ -50,8 +62,11 @@ export const overhead = [
     label: 'Agent and distributor fees',
     id: 'agent_distributor_fees',
     placeholder: '0',
-    tooltip:
-      '<h3>What are agents and distributor fees?</h3><p>An agent is someone who acts on your behalf to sell your product and a distributor is an independent contractor. They take your product and sell them to the customer with an added margin. Both of which will charge fees for their services.</p>',
+    tooltip: {
+      content:
+        '<p>An agent is someone who acts on your behalf to sell your product and a distributor is an independent contractor. They take your product and sell them to the customer with an added margin. Both of which will charge fees for their services.</p>',
+      title: 'What are agents and distributor fees?',
+    },
     type: 'number',
     field: 'overhead_costs',
   },
@@ -59,8 +74,11 @@ export const overhead = [
     label: 'Marketing',
     id: 'marketing',
     placeholder: '0',
-    tooltip:
-      '<h3>What is marketing?</h3><p>This is how you promote your products abroad.</p><p>You can do marketing yourself or pay for marketing services, but either way there will probably be a cost.</p>',
+    tooltip: {
+      content:
+        '<p>This is how you promote your products abroad.</p><p>You can do marketing yourself or pay for marketing services, but either way there will probably be a cost.</p>',
+      title: 'What is marketing?',
+    },
     type: 'number',
     field: 'overhead_costs',
   },
@@ -68,8 +86,11 @@ export const overhead = [
     label: 'Insurance',
     id: 'insurance',
     placeholder: '0',
-    tooltip:
-      '<h3>What is insurance?</h3><p>Like any business activity, exporting has its risks.</p><p>You can minimise these risks by looking at any challenges you might face and deciding whether you should insure against them.</p>',
+    tooltip: {
+      content:
+        '<p>Like any business activity, exporting has its risks.</p><p>You can minimise these risks by looking at any challenges you might face and deciding whether you should insure against them.</p>',
+      title: 'What is insurance?',
+    },
     type: 'number',
     field: 'overhead_costs',
   },
@@ -77,7 +98,7 @@ export const overhead = [
     label: 'Other overhead costs',
     id: 'other_overhead_costs',
     placeholder: '0',
-    tooltip: '',
+    tooltip: {},
     type: 'number',
     field: 'overhead_costs',
   },
@@ -93,8 +114,10 @@ export const costPerUnit = {
   example: {
     header: (cost) => `Your estimate cost per unit is ${cost}`,
     buttonTitle: 'Estimate',
-    content:
-      '<p class="m-b-0">We calculated this by:</p><ul class="list-dot"><li>taking your total overhead costs</li><li>dividing it by the number of units you want to export</li><li>adding this to your direct costs</li></ul><p class="m-v-0">You may want to adjust this estimate, especially if your overhead costs are varied.</p>',
+    content: {
+      content:
+        '<p class="m-b-0">We calculated this by:</p><ul class="list-dot"><li>taking your total overhead costs</li><li>dividing it by the number of units you want to export</li><li>adding this to your direct costs</li></ul><p class="m-v-0">You may want to adjust this estimate, especially if your overhead costs are varied.</p>',
+    },
   },
   field: 'total_cost_and_price',
 }
@@ -112,8 +135,11 @@ export const averagePrice = {
 export const netPrice = {
   label: 'Net Price',
   id: 'net_price',
-  tooltip:
-    '<h3>What is net price?</h3><p>This is the price the customer pays for a single unit of your product before taxes have been added to the final price.</p>',
+  tooltip: {
+    content:
+      '<p>This is the price the customer pays for a single unit of your product before taxes have been added to the final price.</p>',
+    title: 'What is net price?',
+  },
   type: 'number',
   placeholder: '0',
   example: {
@@ -126,8 +152,11 @@ export const netPrice = {
 export const localTaxes = {
   label: 'local Taxes',
   id: 'local_tax_charges',
-  tooltip:
-    "<h3>What are local sales taxes?</h3><p>These are the taxes charged by the UK government, the government of your target market, or both.</p><p>In the UK, our local sales tax is called VAT or goods and services tax (GST).</p><p>The rate of tax you pay depends on where you're exporting.</p><p>This means that if you change target country or export to more than one place, you'll need to recalculate your gross price per unit.</p>",
+  tooltip: {
+    content:
+      "<p>These are the taxes charged by the UK government, the government of your target market, or both.</p><p>In the UK, our local sales tax is called VAT or goods and services tax (GST).</p><p>The rate of tax you pay depends on where you're exporting.</p><p>This means that if you change target country or export to more than one place, you'll need to recalculate your gross price per unit.</p>",
+    title: 'What are local sales taxes?',
+  },
   type: 'number',
   placeholder: '0',
   example: {